# openapi/spec/openapi.yml
openapi: 3.0.3
info:
  title: airdnb API
  version: '1'
servers:
  - url: http://localhost:8080/v1
    description: API
paths:
  /guests/users:
    post:
      tags:
        - guestsUsers
      operationId: guestsPostUsers
      description: 新しいユーザー(ゲスト)を作成
      security:
        - Bearer: []
      requestBody:
        description: 新しいユーザーの作成リクエスト
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestsPostUsersReq'
      responses:
        204:
          description: No Content
  /guests/users/me:
    get:
      tags:
        - guestsUsers
      operationId: guestsGetMe
      description: ユーザーを取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsGetMeRes'
  /guests/rooms:
    get:
      #      TODO: 検索クエリを追加
      tags:
        - guestsRooms
      operationId: guestsGetRooms
      description: 部屋を取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsGetRoomsRes'
  /guests/rooms/{roomId}:
    get:
      tags:
        - guestsRooms
      operationId: guestsGetRoom
      description: 部屋の詳細情報を取得
      security:
        - Bearer: []
      parameters:
        - in: path
          name: roomId
          description: Room ID
          schema:
            type: string
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsRoom'
  /guests/reservations:
    get:
      #      TODO: 時間でcursor
      tags:
        - guestsReservations
      operationId: guestsGetReservations
      description: 予約一覧を取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsGetReservationsRes'
    post:
      tags:
        - guestsReservations
      operationId: postReservations
      description: 新しい予約
      security:
        - Bearer: []
      requestBody:
        description: 新しい予約リクエスト
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostReservationsReq'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsReservation'
  /hosts/users/me:
    get:
      tags:
        - hostsUsers
      operationId: hostsGetMe
      description: ユーザーを取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostsGetMeRes'
  /hosts/rooms:
    get:
      tags:
        - hostsRooms
      operationId: hostsGetRooms
      description: 自分が掲載している部屋を取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostsGetRoomsRes'
    post:
      tags:
        - hostsRooms
      operationId: hostsPostRooms
      description: 新しい部屋を掲載
      security:
        - Bearer: []
      requestBody:
        description: 新しい部屋の掲載リクエスト
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostsPostRoomsReq'
      responses:
        204:
          description: NoContent
  /hosts/rooms/{roomId}:
    put:
      tags:
        - hostsRooms
      operationId: hostsPutRooms
      description: 部屋の情報を変更
      security:
        - Bearer: []
      parameters:
        - in: path
          name: roomId
          description: Room ID
          schema:
            type: string
          required: true
      requestBody:
        description: 新しい部屋の掲載リクエスト
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostsPutRoomsReq'
      responses:
        204:
          description: NoContent
    delete:
      tags:
        - hostsRooms
      operationId: hostsDeleteRooms
      description: 部屋を削除
      security:
        - Bearer: []
      parameters:
        - in: path
          name: roomId
          description: Room ID
          schema:
            type: string
          required: true
      responses:
        204:
          description: NoContent
  /hosts/reservations:
    get:
      # TODO: 時間でcursor
      tags:
        - hostsReservations
      operationId: hostsGetReservations
      description: 予約一覧を取得
      security:
        - Bearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostsGetReservationsRes'

components:
  schemas:
    GuestsPostUsersReq:
      type: object
      required:
        - user_id
        - name
      properties:
        user_id:
          type: string
          example: alskdjfaslkdf
        name:
          type: string
          example: たなか
    GuestsGetMeRes:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: alskdjfaslkdf
        name:
          type: string
          example: たなか
    GuestsGetRoomsRes:
      type: object
      required:
        - rooms
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/GuestsRoom'
    GuestsGetReservationsRes:
      type: object
      required:
        - reservations
      properties:
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/GuestsReservation'
    PostReservationsReq:
      type: object
      required:
        - check_in
        - check_out
        - room_id
      properties:
        check_in:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        check_out:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        room_id:
          type: string
          example: asoiejfl

    GuestsRoom:
      type: object
      required:
        - id
        - name
        - price
        - host_id
      properties:
        id:
          type: string
          example: laksjdfasdf
        name:
          type: string
          example: リッツカールトン
        price:
          type: integer
          example: 2000000
        host_id:
          type: string
          example: llkajsdfil
    GuestsReservation:
      type: object
      required:
        - id
        - check_in
        - check_out
        - room_id
        - room_name
      properties:
        id:
          type: string
          example: lsakdjfoie
        check_in:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        check_out:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        room_id:
          type: string
          example: slafdjio
        room_name:
          type: string
    HostsGetMeRes:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: alskdjfaslkdf
        name:
          type: string
          example: たなか
    HostsGetRoomsRes:
      type: object
      required:
        - rooms
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/HostsRoom'
    HostsPostRoomsReq:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: リッツカールトン
        price:
          type: integer
          example: 300000
    HostsPutRoomsReq:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: リッツカールトン
        price:
          type: integer
          example: 300000
    HostsGetReservationsRes:
      type: object
      required:
        - reservations
      properties:
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/HostsReservation'
    HostsReservation:
      type: object
      required:
        - id
        - check_in
        - check_out
        - room_id
        - room_name
        - guest_id
        - guest_name
      properties:
        id:
          type: string
          example: lsakdjfoie
        check_in:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        check_out:
          type: string
          format: date-time
          example: 2006-01-02T15:04:05Z07:00
        room_id:
          type: string
          example: slafdjio
        room_name:
          type: string
          example: リッツカールトン
        guest_id:
          type: string
          example: laskjfoie
        guest_name:
          type: string
          example: きむら
    HostsRoom:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
          example: laksjdfasdf
        name:
          type: string
          example: リッツカールトン
        price:
          type: integer
          example: 2000000
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Credentials or access token for API